pyproject.toml
[tool.poetry]
name = "gcp-services-poc"
version = "0.1.0"
description = ""
authors = ["Matt Webster <me@matt-webster.net>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.1"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
python-multipart = "^0.0.5"
google-cloud-storage = "^1.38.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
